FinalDeliverables/requirements.txt
APIFlask==1.1.3
apispec==6.0.2
click==8.1.3
Flask==2.1.3
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.0
greenlet==2.0.1
gunicorn==20.1.0
ibm-db==3.1.3
ibm-db-sa==0.3.8
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
python-http-client==3.3.7
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.44
starkbank-ecdsa==2.2.0
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.10.0
